Removes models
