+++ SystemTest/main.go
@@ -18,6 +18,42 @@ const (
 	testUserEmail    = "mohamadkhalafi.82@gmail.com"
 )
 
+
+type Item struct {
+	Id          int    `json:"id"`
+	Title       string `json:"title"`
+	CategoryId  int    `json:"category_id"`
+	Price       string `json:"price"`
+	Description string `json:"description"`
+	ImageIds    []int  `json:"image_ids"`
+}
+
+type Banner struct {
+	ImageId int `json:"image_id"`
+	Order   int `json:"order"`
+}
+
+type CreateItemRequest struct {
+	Title       string   `json:"title"`
+	Category    int      `json:"category"`
+	Price       int      `json:"price"`
+	Description string   `json:"description"`
+	Banners     []Banner `json:"banners"`
+}
+
+type GetAllItemsResponse struct {
+	Count   int `json:"count"`
+	Results struct {
+		Items    []Item `json:"items"`
+		MaxPrice int    `json:"max_price"`
+	} `json:"results"`
+}
+
+type LoginRequest struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
+
 type ServiceClient struct {
 	c *http.Client
 }
@@ -171,4 +207,8 @@ func (s *ServiceClient) Login() (string, error) {
 	}
 
 	return get.Cookies()[0].Value, nil
+}
+
+func createUrl(route string) string {
+	return fmt.Sprintf("%s/%s", BaseUrl, route)
 }
