--- SystemTest/utils.go
@@ -1,7 +1,0 @@
-package main
-
-import "fmt"
-
-func createUrl(route string) string {
-	return fmt.Sprintf("%s/%s", BaseUrl, route)
-}
